Fix of weather service methods' returning types (now returns object instead of it's string representation) and moved Info classes into weatherservice package
